0.3.4: Fix unknown key awaitInput call timing
If you just press enter, keypress event will be emitted once.
However, if you press 'aaaa' multiple times, the keypress event will be emitted
five times: 4 times for the 'a's, and one time for the enter key. This causes a problem,
because on an unknown key, we synchronously re-call the awaitInput function, and
synchronously re-attach ourselves to the input, causing the next tick
to fire the event handler.

setTimeout gives the keypress event to emit all the rest of the key events,
and then once it's drained, start listening to further key presses.